move About Window to Preferences Window
