+++ Shifty.xcworkspace/xcuserdata/Nate.xcuserdatad/xcdebugger/Breakpoints_v2.xcbkptlist
@@ -10,11 +10,11 @@
             ignoreCount = "0"
             continueAfterRunningActions = "No"
             filePath = "Shifty/StatusMenuController.swift"
-            timestampString = "532070618.138051"
+            timestampString = "532074700.085614"
             startingColumnNumber = "9223372036854775807"
             endingColumnNumber = "9223372036854775807"
-            startingLineNumber = "206"
-            endingLineNumber = "206"
+            startingLineNumber = "204"
+            endingLineNumber = "204"
             landmarkName = "setToSchedule()"
             landmarkType = "7">
          </BreakpointContent>
