+++ Shifty/MainMenu.xib
@@ -1,8 +1,8 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<document type="com.apple.InterfaceBuilder3.Cocoa.XIB" version="3.0" toolsVersion="13196" targetRuntime="MacOSX.Cocoa" propertyAccessControl="none" useAutolayout="YES" customObjectInstantitationMethod="direct">
+<document type="com.apple.InterfaceBuilder3.Cocoa.XIB" version="3.0" toolsVersion="13529" targetRuntime="MacOSX.Cocoa" propertyAccessControl="none" useAutolayout="YES" customObjectInstantitationMethod="direct">
     <dependencies>
         <deployment identifier="macosx"/>
-        <plugIn identifier="com.apple.InterfaceBuilder.CocoaPlugin" version="13196"/>
+        <plugIn identifier="com.apple.InterfaceBuilder.CocoaPlugin" version="13529"/>
         <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
     </dependencies>
     <objects>
@@ -68,12 +68,6 @@
                     </connections>
                 </menuItem>
                 <menuItem isSeparatorItem="YES" id="OE0-tz-FKo"/>
-                <menuItem title="About..." id="B5o-hm-scl">
-                    <modifierMask key="keyEquivalentModifierMask"/>
-                    <connections>
-                        <action selector="aboutClicked:" target="YHE-ZD-Fjn" id="Rhu-71-Pgb"/>
-                    </connections>
-                </menuItem>
                 <menuItem title="Preferences..." id="GJi-nB-WzR">
                     <modifierMask key="keyEquivalentModifierMask"/>
                     <connections>
